Improve detection of geometry-containing elements
Elements whose name contains geometry, position or the_geom would be detected
as geometry elements and therefore not shown. In the case of e.g. gml:TimePeriod
with beginPosition and endPosition then this logic is insufficient. A
match on the whole tag name is also insufficient, as names such as e.g.
geometry2d would not be recognised as geometries.

Refs #5